+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### FEATURES/ENHANCEMENTS:
 
 * DNS
-  * Added the `MultiProviderDnssec` struct and field on `ZoneCreate` and `ZoneResponse` to support multi-signer DNSSEC configuration on Edge DNS zones.
+  * Added the `MultiProviderDNSSEC` struct and field on `ZoneCreate` and `ZoneResponse` to support multi-signer DNSSEC configuration on Edge DNS zones.
 
 ## 13.3.0 (Jun 29, 2026)